Fix Model
