+++ model/comment.js
@@ -1,10 +1,10 @@
 const mongoose = require("mongoose")
 
 const CommentSchema = mongoose.Schema({
-    detail : {type: String},
-    date: {type: String},
-    author: { type: mongoose.Schema.Types.ObjectId, ref: 'user' },
-    post: { type: mongoose.Schema.Types.ObjectId, ref: 'post' }
+    content : {type: String},
+    createByDate: {type: String},
+    author: { type: Object, ref: 'user' },
+    post: { type: Object, ref: 'article' }
 })
 const Comment = mongoose.model('comments', CommentSchema,'comments');
 module.exports = Comment;
